+++ packages/components/image-upload/src/index.ts
@@ -3,5 +3,6 @@
 export { ImageUpload, type ImageUploadProps } from "./image-upload";
 export {
   MultiImageUpload,
-  type MultiUploadImgState as FileState,
+  type MultiImageUploadProps,
+  type MultiUploadImgState,
 } from "./multi-image-upload";